bismillah final

+++ app/controllers/Skema.php
@@ -23,11 +23,16 @@ class Skema extends Controller {
     exit;
   }
 
-  public function list() {
+  public function list($user_type = null) {
 
     // check login
     if (isset($_SESSION['user-type'])) {
       $this->model('User_model')->checkUserLogin($_SESSION['user-type']);
+    }
+
+    if ($_SESSION['user-type'] == "asesor" && $user_type != "asesor") {
+      header('Location: ' . BASEURL . '/skema/list/asesor');
+      exit;
     }
 
     $data['user-type'] = $_SESSION['user-type'];
@@ -41,13 +46,30 @@ class Skema extends Controller {
 
     $data['page-link'] = $this->breadcrumbs;
 
-
     if (isset($_POST['search-key']) && $_POST['search-key'] != "") {
       $data["list-skema"] = $this->model("Skema_model")->searchDataSkema($_POST['search-key']);
     } else if (isset($_POST['kategori'])) {
       $data["list-skema"] = $this->model("Skema_model")->filterSkemaByKategori($_POST);
     } else {
-      $data["list-skema"] = ($data['user-type'] == "admin") ? $this->model("Skema_model")->fetchAllSkema() : ($data['user-type'] == "asesi" ? $this->model("Skema_model")->getSchemaOfAsesi($bio_id) : $this->model("Skema_model")->getSchemaOfAsesor($bio_id));
+      $data["list-skema"] = $this->model("Skema_model")->fetchAllSkema();
+    }
+
+    if ($_SESSION['user-type'] == "asesi") {
+      $id_jurusan_asesi = $this->model("User_model")->getIdJurusanOfAsesi($bio_id);
+
+      if (isset($_POST['search-key']) && $_POST['search-key'] != "") {
+        $data["list-skema"] = $this->model("Skema_model")->searchDataSkemaOfAsesi($_POST['search-key'], $id_jurusan_asesi);
+      } else if (isset($_POST['kategori'])) {
+        $data["list-skema"] = $this->model("Skema_model")->filterSkemaByKategoriOfJurusan($_POST, $id_jurusan_asesi);
+      } else {
+        $data['list-skema'] = $this->model("Skema_model")->getSchemaOfJurusan($id_jurusan_asesi);
+      }
+    }
+
+    if ($_SESSION['user-type'] == "asesi" && $user_type == "asesi") {
+      $data["list-skema"] = $this->model("Skema_model")->getSchemaOfAsesi($bio_id);
+    } else if ($_SESSION['user-type'] == "asesor" && $user_type == "asesor") {
+      $data["list-skema"] = $this->model("Skema_model")->getSchemaOfAsesor($bio_id);
     }
 
     $this->view('templates/sidebar');
@@ -55,9 +77,6 @@ class Skema extends Controller {
     if ($data['user-type'] == "admin") {
       $data['list-skkni'] = $this->model('Skema_model')->fetchAllSKKNI();
       $data['list-asesor'] = $this->model('Asesor_model')->fetchAllAsesor();
-    }
-
-    if ($data['user-type'] != "asesi") {
       $data['list-jurusan'] = $this->model('User_model')->fetchAllJurusan();
     }
 
@@ -70,6 +89,8 @@ class Skema extends Controller {
   }
 
   public function add() {
+    $this->model("User_model")->checkUserLogin("admin");
+
     if ($this->model("Skema_model")->addSkema($_POST) > 0) {
       Flasher::setFlash("Skema sertifikasi baru berhasil ditambahkan", "success");
     }
@@ -77,31 +98,14 @@ class Skema extends Controller {
     exit;
   }
 
-  public function create() {
-    $data['page-title'] = "Skema";
-    $data['skkni'] = $this->model("Skema_model")->fetchAllSKKNI();
-    $data['jurusan'] = $this->model("Skema_model")->fetchAllJurusan();
-    $data['asesor'] = $this->model("Skema_model")->fetchAllAsesor();
-    $this->view('templates/header', $data);
-    $this->view('templates/navbar/main-navbar');
-    $this->view('skema/create', $data);
-    $this->view('templates/footer');
-  }
+  public function update($idSkema = null) {
+    $this->model("User_model")->checkUserLogin("admin");
 
-  public function status($id, $status) {
-    $newStatus = $status == "Aktif" ? "Nonaktif" : "Aktif";
-    if ($this->model("Skema_model")->updateStatus($id, $newStatus) > 0) {
-      header('Location: ' . BASEURL . '/skema/index');
-    } else {
-      Flasher::setFlash("Harap isi dengan benar", 'warning');
-      header('Location: ' . BASEURL . '/skema/index');
+    if ($idSkema == null || $this->model("Skema_model")->isSkemaExist($idSkema) == 0) {
+      header('Location: ' . BASEURL . '/dashboard');
+      exit;
     }
-  }
 
-  public function update($idSkema) {
-    // $nama = preg_replace("/(?<=[a-z])(?=[A-Z])/", " ", $nama);
-    // $level = preg_replace("/(?<=[a-z])(?=[A-Z])/", " ", $level);
-    // $idSkema = $this->model("Persyaratan_model")->getIdSkemaByNama($nama, $level)['id'];
     $level = $this->model("Skema_model")->getSingleSkema($idSkema)['level'];
     $syaratExisting = [];
     $tempCheck = [];
@@ -127,30 +131,92 @@ class Skema extends Controller {
     // var_dump($tempCheck);
     // die;
     if ($this->model("Skema_model")->updateSkema($idSkema, $_POST) > 0 || $_POST['level'] == $level) {
-      header('Location: ' . BASEURL . '/skema/index');
+      Flasher::setFlash("Skema sertifikasi berhasil diperbaharui", "success");
     } else {
       Flasher::setFlash("Harap ganti level dari skemanya", 'warning');
-      header('Location: ' . BASEURL . '/skema/detail/' . $idSkema);
     }
+    header('Location: ' . BASEURL . '/skema/detail/' . $idSkema);
+    exit;
   }
 
-  public function detail($idSkema) {
+  public function detail($idSkema = null) {
+    $this->model("User_model")->checkUserLogin($_SESSION['user-type']);
+
+    if ($idSkema == null || $this->model("Skema_model")->isSkemaExist($idSkema) == 0) {
+      header('Location: ' . BASEURL . '/dashboard');
+      exit;
+    }
+
+    $data["data-skema"] = $this->model("Skema_model")->getSingleSkema($idSkema);
+
+    $data['user-type'] = $_SESSION['user-type'];
+
+    $username = $_SESSION['username'];
+    $id_bio = $this->model("User_model")->getIdBioByUsername($data['user-type'], $username);
+
+    if ($data['user-type'] == "asesi") {
+      $id_jurusan_asesi = $this->model("User_model")->getIdJurusanOfAsesi($id_bio);
+      if ($data['data-skema']['id_jurusan'] != $id_jurusan_asesi) {
+        header('Location: ' . BASEURL . '/dashboard');
+        exit;
+      }
+      $data['status-pendaftaran'] = $this->model("Asesi_model")->asesiSudahDaftar($id_bio, $idSkema);
+    } else if ($data['user-type'] == "asesor") {
+      $skema_asesor = $this->model('Skema_model')->getSchemaOfAsesor($id_bio);
+      $all_id_skema_asesor = array_map(function ($v) {
+        return $v['id'];
+      }, $skema_asesor);
+      if (!in_array($idSkema, $all_id_skema_asesor)) {
+        header('Location: ' . BASEURL . '/dashboard');
+        exit;
+      }
+    }
+
+    if ($data['user-type'] == "asesor") {
+      require_once '../app/controllers/Asesor.php';
+    }
+
     $data['page-title'] = "Detail Skema Sertifikasi";
     $data['id-skema'] = $idSkema;
     $data['list-level'] = $this->model("Persyaratan_model")->fetchLevelBySkema($idSkema);
     $data['persyaratan-skema'] = $this->model("Persyaratan_model")->fetchPersyaratanBySkema($idSkema);
-    // if ($nama != "") {
-    //   $data['skema-selected'] = preg_replace("/(?<=[a-z])(?=[A-Z])/", " ", $nama);
-    //   if ($level != "") {
-    //     $data['level-selected'] = preg_replace("/(?<=[a-z])(?=[A-Z])/", " ", $level);
-    //   }
-    // }
-    $data["data-skema"] = $this->model("Skema_model")->getSingleSkema($idSkema);
+
     $data['list-persyaratan-umum'] = $this->model("Persyaratan_model")->fetchAllPersyaratan("Umum");
     $data['list-persyaratan-teknis'] = $this->model("Persyaratan_model")->fetchAllPersyaratan("Teknis");
     $data["skkni"] = $this->model("Skema_model")->fetchAllSKKNI();
     $data['jurusan'] = $this->model("Skema_model")->fetchAllJurusan();
     $data['asesor'] = $this->model("Skema_model")->fetchAllAsesor();
+
+    $user_type_obj = $_SESSION['user-type'];
+
+    $last_activity_user = $_SESSION['last-activity'][$user_type_obj];
+
+
+    $nama_skema = $data['data-skema']['nama_skema'];
+    // var_dump($data['data-skema']);
+
+    $data['id-skema'] = $idSkema;
+    $data['nama-skema'] = $nama_skema;
+    $data['skkni-skema'] = $data['data-skema']['skkni'];
+    $data['asesor-skema'] = $data['data-skema']['asesor'];
+    $data['jurusan-skema'] = $data['data-skema']['jurusan'];
+    $data['level-skema'] = $data['data-skema']['level'];
+    $data['status-skema'] = $data['data-skema']['status'];
+
+    if ($last_activity_user->lastChange == "first" && $last_activity_user->first->name != $nama_skema && $last_activity_user->first->id != $idSkema) {
+      $last_activity_user->second->name = $nama_skema;
+      $last_activity_user->second->id = $idSkema;
+      $last_activity_user->lastChange = "second";
+    } else if ($last_activity_user->lastChange == "second" && $last_activity_user->second->name != $nama_skema && $last_activity_user->second->id != $idSkema) {
+      $last_activity_user->first->name = $nama_skema;
+      $last_activity_user->first->id = $idSkema;
+      $last_activity_user->lastChange = "first";
+    }
+
+    $_SESSION['last-activity'][$user_type_obj] = $last_activity_user;
+    // var_dump($_SESSION['last-activity']);
+
+    setcookie('last-activity', json_encode($_SESSION['last-activity']), time() + (86400 * 30), "/");
 
     // $idSkema = $this->model("Persyaratan_model")->getIdSkemaByNama($data['skema-selected'], $data['level-selected'])['id'];
     $data['list-kompetensi'] = $this->model("Skema_model")->getKompetensiByIdSkema($idSkema);
@@ -163,7 +229,7 @@ class Skema extends Controller {
     array_push($data['page-link']['link'], "skema/list");
 
     $this->view('templates/sidebar');
-    $this->view('skema/admin/detail', $data);
+    $this->view('skema/' . $data['user-type'] . '/detail', $data);
     $this->view('templates/close_html_tag');
   }
 
@@ -180,8 +246,28 @@ class Skema extends Controller {
     echo json_encode($dataPersyaratan);
   }
 
+  public function delete($idSkema = null) {
+    $this->model("User_model")->checkUserLogin("admin");
+
+    if ($idSkema == null || $this->model("Skema_model")->isSkemaExist($idSkema) == 0) {
+      header('Location: ' . BASEURL . '/dashboard');
+      exit;
+    }
+
+    if ($this->model("Skema_model")->deleteSkemaById($idSkema) > 0) {
+      Flasher::setFlash("Skema sertifikasi berhasil dihapus", "success");
+    } else {
+      Flasher::setFlash("Skema sertifikasi tidak ada", "error");
+    }
+
+    header("Location: " . BASEURL . "/skema/list");
+    exit;
+  }
+
   // Unit Kompetensi
   public function asesmen($page = 1) {
+    $this->model("User_model")->checkUserLogin("admin");
+
     $data['page-title'] = "Jadwal Asesmen";
     $data['page'] = $page;
     $data['page-total'] = ceil($this->model("Skema_model")->getTotalAsesmen() / 5);
@@ -192,8 +278,7 @@ class Skema extends Controller {
     }
 
     $data['page-link'] = $this->breadcrumbs;
-    // var_dump($data['list-kompetensi']);
-    // die;
+
     $this->view('templates/header', $data);
     $this->view("templates/sidebar");
 
@@ -202,73 +287,104 @@ class Skema extends Controller {
     $this->view('templates/close_html_tag');
   }
 
-  public function tambah_kompetensi($idSkema) {
-    $data['page-title'] = "Skema";
+  public function tambah_kompetensi($idSkema = null) {
+    $this->model("User_model")->checkUserLogin("admin");
+
+    if ($idSkema == null || $this->model("Skema_model")->isSkemaExist($idSkema) == 0) {
+      header('Location: ' . BASEURL . '/dashboard');
+      exit;
+    }
+
     $data['id-skema'] = $idSkema;
     // $data['nama-skema'] = preg_replace("/(?<=[a-z])(?=[A-Z])/", " ", $nama);
     // $data['level-skema'] = preg_replace("/(?<=[a-z])(?=[A-Z])/", " ", $level);
-    if (isset($_POST['tambah'])) {
-      // $idSkema = $this->model("Persyaratan_model")->getIdSkemaByNama($data['nama-skema'], $data['level-skema'])['id'];
-      $nama_skema = $this->model("Skema_model")->getSingleSkema($idSkema)['nama_skema'];
-      $level = $this->model("Skema_model")->getSingleSkema($idSkema)['level'];
-      if ($_FILES['soal-kompetensi']['error'] !== 4) {
-        $_POST['soal-kompetensi'] = $this->checkExtensionOfFile($nama_skema, $level, BASEURL . '/skema/tambah_kompetensi/' . $data['id-skema']);
-        if (!$_POST['soal-kompetensi']) {
-          $_POST['soal-kompetensi'] = $data['list-kompetensi']['file_opsional'];
-          header('Location: ' . BASEURL . '/skema/tambah_kompetensi/' . $data['id-skema']);
-          return;
-        }
-      }
-      if ($this->model("Skema_model")->addKompetensi($idSkema, $_POST) > 0) {
-        header('Location: ' . BASEURL . '/skema/detail/' . $data['id-skema']);
-      } else {
-        Flasher::setFlash("Nama kompetensi sudah ada", 'warning');
+    // $idSkema = $this->model("Persyaratan_model")->getIdSkemaByNama($data['nama-skema'], $data['level-skema'])['id'];
+    $nama_skema = $this->model("Skema_model")->getSingleSkema($idSkema)['nama_skema'];
+    $level = $this->model("Skema_model")->getSingleSkema($idSkema)['level'];
+    if ($_FILES['soal-kompetensi']['error'] !== 4) {
+      $_POST['soal-kompetensi'] = $this->checkExtensionOfFile($nama_skema, $level, BASEURL . '/skema/tambah_kompetensi/' . $data['id-skema']);
+      if (!$_POST['soal-kompetensi']) {
+        $_POST['soal-kompetensi'] = $data['list-kompetensi']['file_opsional'];
         header('Location: ' . BASEURL . '/skema/tambah_kompetensi/' . $data['id-skema']);
       }
+    }
+    if ($this->model("Skema_model")->addKompetensi($idSkema, $_POST) > 0) {
+      Flasher::setFlash("Kompetensi baru berhasil ditambahkan", "success");
+      header('Location: ' . BASEURL . '/skema/detail/' . $data['id-skema']);
     } else {
-      $this->view('templates/header', $data);
-      $this->view('templates/navbar/main-navbar');
-      $this->view('kompetensi/tambah', $data);
-      $this->view('templates/footer');
+      Flasher::setFlash("Nama kompetensi sudah ada", 'warning');
+      header('Location: ' . BASEURL . '/skema/tambah_kompetensi/' . $data['id-skema']);
+    }
+    exit;
+  }
+
+  public function edit_kompetensi($id = null, $idSkema = null) {
+    $this->model("User_model")->checkUserLogin("admin");
+
+    if ($idSkema == null || $id == null || $this->model("Skema_model")->isSkemaExist($idSkema) == 0) {
+      header('Location: ' . BASEURL . '/dashboard');
+      exit;
+    }
+
+    $data['page-title'] = "Edit Kompetensi";
+    $data['id-kompetensi'] = $id;
+    $data['id-skema'] = $idSkema;
+
+    $data['page-link'] = $this->breadcrumbs;
+    array_push($data['page-link']['name'], "List Skema Sertifikasi");
+    array_push($data['page-link']['link'], "skema/list");
+
+    array_push($data['page-link']['name'], 'Detail Skema Sertifikasi');
+    array_push($data['page-link']['link'], 'skema/detail/' . $idSkema);
+
+    $data['list-kompetensi'] = $this->model("Skema_model")->getKompetensiByIdKompetensi($id);
+
+    $this->view('templates/header', $data);
+    $this->view('templates/sidebar');
+    $this->view('kompetensi/edit', $data);
+    $this->view('templates/close_html_tag');
+    // $data['nama-skema'] = preg_replace("/(?<=[a-z])(?=[A-Z])/", " ", $nama);
+    // $data['level-skema'] = preg_replace("/(?<=[a-z])(?=[A-Z])/", " ", $level);
+    if (isset($_POST['ubah'])) {
     }
   }
 
-  public function edit_kompetensi($id, $idSkema) {
-    $data['page-title'] = "Skema";
-    $data['id-kompetensi'] = $id;
-    $data['id-skema'] = $idSkema;
-    // $data['nama-skema'] = preg_replace("/(?<=[a-z])(?=[A-Z])/", " ", $nama);
-    // $data['level-skema'] = preg_replace("/(?<=[a-z])(?=[A-Z])/", " ", $level);
+  public function edit_asesmen_post($id = null, $idSkema = null) {
+    $this->model("User_model")->checkUserLogin("admin");
+
+    if ($idSkema == null || $id == null || $this->model("Skema_model")->isSkemaExist($idSkema) == 0) {
+      header('Location: ' . BASEURL . '/dashboard');
+      exit;
+    }
+
     $data['list-kompetensi'] = $this->model("Skema_model")->getKompetensiByIdKompetensi($id);
-    if (isset($_POST['ubah'])) {
-      $nama_skema = $this->model("Skema_model")->getSingleSkema($idSkema)['nama_skema'];
-      $level = $this->model("Skema_model")->getSingleSkema($idSkema)['level'];
-      if ($_FILES['soal-kompetensi']['error'] !== 4) {
-        $_POST['soal-kompetensi'] = $this->checkExtensionOfFile($nama_skema, $level, BASEURL . '/skema/edit_kompetensi/' . $id . '/' . $data['id-skema']);
-        // var_dump($_POST['soal-kompetensi']);
-        // die;
-        if (!$_POST['soal-kompetensi']) {
-          $_POST['soal-kompetensi'] = $data['list-kompetensi']['file_opsional'];
-          header('Location: ' . BASEURL . '/skema/edit_kompetensi/' . $id . '/' . $data['id-skema']);
-          return;
-        }
-      } else {
-        $_POST['soal-kompetensi'] = $data['list-kompetensi']['file_opsional'];
-      }
-      // var_dump($_POST);
+    $data['id-skema'] = $idSkema;
+
+
+    $nama_skema = $this->model("Skema_model")->getSingleSkema($idSkema)['nama_skema'];
+    $level = $this->model("Skema_model")->getSingleSkema($idSkema)['level'];
+    if ($_FILES['soal-kompetensi']['error'] !== 4) {
+      $_POST['soal-kompetensi'] = $this->checkExtensionOfFile($nama_skema, $level, BASEURL . '/skema/edit_kompetensi/' . $id . '/' . $data['id-skema']);
+      // var_dump($_POST['soal-kompetensi']);
       // die;
-      if ($this->model("Skema_model")->updateKompetensi($id, $_POST) > 0 || $_POST['nama-kompetensi'] == $data['list-kompetensi']['nama_kompetensi']) {
-        header('Location: ' . BASEURL . '/skema/detail/' . $data['id-skema']);
-      } else {
-        Flasher::setFlash("Nama kompetensi sudah ada", 'warning');
+      if (!$_POST['soal-kompetensi']) {
+        $_POST['soal-kompetensi'] = $data['list-kompetensi']['file_opsional'];
         header('Location: ' . BASEURL . '/skema/edit_kompetensi/' . $id . '/' . $data['id-skema']);
+        return;
       }
     } else {
-      $this->view('templates/header', $data);
-      $this->view('templates/navbar/main-navbar');
-      $this->view('kompetensi/ubah', $data);
-      $this->view('templates/footer');
+      $_POST['soal-kompetensi'] = $data['list-kompetensi']['file_opsional'];
     }
+    // var_dump($_POST);
+    // die;
+    if ($this->model("Skema_model")->updateKompetensi($id, $_POST) > 0 || $_POST['nama-kompetensi'] == $data['list-kompetensi']['nama_kompetensi']) {
+      Flasher::setFlash("Kompetensi berhasil diubah", "success");
+      header('Location: ' . BASEURL . '/skema/detail/' . $data['id-skema']);
+    } else {
+      Flasher::setFlash("Nama kompetensi sudah ada", 'warning');
+      header('Location: ' . BASEURL . '/skema/edit_kompetensi/' . $id . '/' . $data['id-skema']);
+    }
+    exit;
   }
 
   private function checkExtensionOfFile($nama, $level, $url) {
@@ -309,15 +425,25 @@ class Skema extends Controller {
     return $namaFileBaru;
   }
 
-  public function hapus_kompetensi($id, $idSkema) {
+  public function hapus_kompetensi($id = null, $idSkema = null) {
+    $this->model("User_model")->checkUserLogin("admin");
+
+    if ($idSkema == null || $this->model("Skema_model")->isKompetensiExist($id) == 0 || $this->model("Skema_model")->isSkemaExist($idSkema) == 0) {
+      header('Location: ' . BASEURL . '/dashboard');
+      exit;
+    }
     // $nama = preg_replace("/(?<=[a-z])(?=[A-Z])/", " ", $nama);
     // $level = preg_replace("/(?<=[a-z])(?=[A-Z])/", " ", $level);
+
+
+
     if ($this->model("Skema_model")->deleteKompetensiById($id) > 0) {
-      header('Location: ' . BASEURL . "/skema/detail/$idSkema");
+      Flasher::setFlash("Unit kompetensi berhasil dihapus", 'success');
     } else {
       Flasher::setFlash("Ada yg salah dari hapusnya", 'warning');
-      header('Location: ' . BASEURL . "/skema/detail/$idSkema");
     }
+    header('Location: ' . BASEURL . "/skema/detail/$idSkema");
+    exit;
   }
 
   public function countRegistered($id) {
